requirements.txt
google-genai==1.12.1
dotenv==1.1.0